Helpers fixed
