+++ lib/siwapp_web/live/customer_live/edit.ex
@@ -1,8 +1,6 @@
 defmodule SiwappWeb.CustomersLive.Edit do
   use SiwappWeb, :live_view
 
-  import SiwappWeb.HelpersLive
-
   alias Siwapp.Customers
 
   def mount(_params, _session, socket) do
