+++ lib/siwapp_web/live/invoices_live/invoices.ex
@@ -1,0 +1,3 @@
+defmodule SiwappWeb.InvoicesLive.Index do
+  use SiwappWeb, :live_view
+end
